+++ java_tools/configuration_definition/src/main/java/com/rusefi/newparse/outputs/SdLogWriter.java
@@ -28,15 +28,17 @@ public class SdLogWriter {
         ps.close();
     }
 
-    public void writeSdLogs(ParseState parser, String sourceName, String category) {
+    // baseOffset is where this struct starts within the whole output channel space, so each field's
+    // absolute offset is baseOffset + its offset within the struct.
+    public void writeSdLogs(ParseState parser, int baseOffset, String category) {
         // Assume the last struct is the one we want...
         Struct s = parser.getStructs().get(parser.getStructs().size() - 1);
 
         StructLayout sl = new StructLayout(0, "root", s);
 
-        SdLogVisitor v = new SdLogVisitor(sourceName, category);
+        SdLogVisitor v = new SdLogVisitor(category);
         StructNamePrefixer prefixer = new StructNamePrefixer('.');
 
-        v.visit(sl, ps, prefixer, 0, new int[0]);
+        v.visit(sl, ps, prefixer, baseOffset, new int[0]);
     }
 }